+++ src/main/java/com/rcextract/minecord/event/ChannelSetNameEvent.java
@@ -4,14 +4,14 @@ import org.bukkit.event.HandlerList;
 
 import com.rcextract.minecord.Channel;
 
-public class ChannelRenameEvent extends ChannelEvent {
+public class ChannelSetNameEvent extends ChannelEvent {
 
 	private static final HandlerList handlers = new HandlerList();
 	public static HandlerList getHandlerList() {
 		return handlers;
 	}
 	private String name;
-	public ChannelRenameEvent(Channel channel, String name) {
+	public ChannelSetNameEvent(Channel channel, String name) {
 		super(channel);
 		this.name = name;
 	}