+++ src/main/java/com/rcextract/minecord/event/ServerSetNameEvent.java
@@ -1,0 +1,30 @@
+package com.rcextract.minecord.event;
+
+import org.bukkit.event.HandlerList;
+
+import com.rcextract.minecord.Server;
+
+public class ServerSetNameEvent extends ServerEvent {
+
+	private static final HandlerList handlers = new HandlerList();
+	public static HandlerList getHandlerList() {
+		return handlers;
+	}
+	private String name;
+	public ServerSetNameEvent(Server server, String name) {
+		super(server);
+		this.name = name;
+	}
+
+	public String getName() {
+		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
+	}
+	@Override
+	public HandlerList getHandlers() {
+		return handlers;
+	}
+
+}